Rename list.rs to lists.rs
